se implementaron los grupos, envio de mensajes a grupos y carga de mensajes  de grupos, TODO : implementar carga de chats y grupos concurrente

+++ AppChat/src/main/java/umu/tds/AppChat/dao/MensajeDAO.java
@@ -69,7 +69,7 @@ public class MensajeDAO implements InterfaceMensajeDAO {
 		Optional<String> messageMsg = message.isBlank() ? Optional.empty() : Optional.of(message);
 		Optional<Integer> emoojiMSG = emoji.isBlank() ? Optional.empty() : Optional.of(Integer.parseInt(emoji));
 		
-		ModelMessage msg = new ModelMessage(icon, senderName, sendDate, Integer.parseInt(sender), Integer.parseInt(reciver), messageMsg, emoojiMSG);
+		ModelMessage msg = new ModelMessage(icon, senderName, sendDate, Integer.parseInt(sender), Long.parseLong(reciver), messageMsg, emoojiMSG);
 		return msg;
 	}
